+++ fem/CMakeLists.txt
@@ -24,6 +24,7 @@ set(SRCS
   fespace.cpp
   fespace_ext.cpp
   geom.cpp
+  geom_ext.cpp
   gridfunc.cpp
   hybridization.cpp
   intrules.cpp
@@ -50,6 +51,7 @@ set(HDRS
   fespace.hpp
   fespace_ext.hpp
   geom.hpp
+  geom_ext.hpp
   gridfunc.hpp
   hybridization.hpp
   intrules.hpp
