+++ fem/geom_ext.cpp
@@ -9,108 +9,18 @@
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
 
-#include "../../config/config.hpp"
+#include "../config/config.hpp"
+#include "../general/okina.hpp"
 
-#include "../../general/okina.hpp"
-#include "../../linalg/kernels/vector.hpp"
-
-#include "geometry.hpp"
-#include "../fem.hpp"
-#include "../doftoquad.hpp"
+#include "fem.hpp"
+#include "geom_ext.hpp"
+#include "kernels/geom.hpp"
+#include "../linalg/kernels/vector.hpp"
 
 namespace mfem
 {
-namespace kernels
-{
-namespace geometry
-{
 
-// *****************************************************************************
-template<const int NUM_DOFS_1D,
-         const int NUM_QUAD_1D>
-void Geom2D(const int,const double*,const double*,double*,double*,double*);
-
-// *****************************************************************************
-template<const int NUM_DOFS_1D,
-         const int NUM_QUAD_1D>
-void Geom3D(const int,const double*,const double*,double*,double*,double*);
-
-// *****************************************************************************
-typedef void (*fIniGeom)(const int,const double*,const double*,
-                         double*, double*, double*);
-
-// *****************************************************************************
-static void Geom(const int DIM,
-                 const int NUM_DOFS,
-                 const int NUM_QUAD,
-                 const int numElements,
-                 const double* dofToQuadD,
-                 const double* nodes,
-                 double* J,
-                 double* invJ,
-                 double* detJ)
-{
-   const unsigned int dofs1D = IROOT(DIM,NUM_DOFS);
-   const unsigned int quad1D = IROOT(DIM,NUM_QUAD);
-   const unsigned int id = (DIM<<8)|(dofs1D-2)<<4|(quad1D-2);
-   assert(LOG2(DIM)<=4);
-   assert(LOG2(dofs1D-2)<=4);
-   assert(LOG2(quad1D-2)<=4);
-   static std::unordered_map<unsigned int, fIniGeom> call =
-   {
-      // 2D
-      {0x200,&Geom2D<2,2>},
-      {0x201,&Geom2D<2,3>},
-      {0x202,&Geom2D<2,4>},
-      {0x203,&Geom2D<2,5>},
-      {0x204,&Geom2D<2,6>},
-      {0x205,&Geom2D<2,7>},
-      {0x206,&Geom2D<2,8>},
-      {0x207,&Geom2D<2,9>},
-      {0x210,&Geom2D<3,2>},/*
-      {0x208,&Geom2D<2,10>},
-      {0x209,&Geom2D<2,11>},
-      {0x20A,&Geom2D<2,12>},
-      {0x20B,&Geom2D<2,13>},
-      {0x20C,&Geom2D<2,14>},
-      {0x20D,&Geom2D<2,15>},
-      {0x20E,&Geom2D<2,16>},
-      {0x20F,&Geom2D<2,17>},*/
-      // 3D
-      {0x300,&Geom3D<2,2>},
-      {0x301,&Geom3D<2,3>},
-      {0x302,&Geom3D<2,4>},
-      {0x303,&Geom3D<2,5>},
-      {0x304,&Geom3D<2,6>},
-      {0x305,&Geom3D<2,7>},
-      {0x306,&Geom3D<2,8>},
-      {0x307,&Geom3D<2,9>},
-      {0x321,&Geom3D<4,3>},/*
-      {0x308,&Geom3D<2,10>},
-      {0x309,&Geom3D<2,11>},
-      {0x30A,&Geom3D<2,12>},
-      {0x30B,&Geom3D<2,13>},
-      {0x30C,&Geom3D<2,14>},
-      {0x30D,&Geom3D<2,15>},
-      {0x30E,&Geom3D<2,16>},
-      {0x30F,&Geom3D<2,17>},*/
-   };
-   if (!call[id])
-   {
-      printf("\n[Geom] id \033[33m0x%X\033[m ",id);
-      fflush(stdout);
-   }
-   assert(call[id]);
-   GET_CONST_PTR(dofToQuadD);
-   GET_CONST_PTR(nodes);
-   GET_PTR(J);
-   GET_PTR(invJ);
-   GET_PTR(detJ);
-   call[id](numElements, d_dofToQuadD, d_nodes, d_J, d_invJ, d_detJ);
-}
-
-// *****************************************************************************
-static Geometry *geom = NULL;
+static GeometryExtension *geom = NULL;
 
 // *****************************************************************************
 static void GeomFill(const int dims,
@@ -174,9 +84,9 @@ static void NodeCopyByVDim(const int elements,
 
 
 // *****************************************************************************
-Geometry* Geometry::Get(const FiniteElementSpace& fes,
-                        const IntegrationRule& ir,
-                        const Vector& Sx)
+GeometryExtension* GeometryExtension::Get(const FiniteElementSpace& fes,
+                                          const IntegrationRule& ir,
+                                          const Vector& Sx)
 {
    const Mesh *mesh = fes.GetMesh();
    const GridFunction *nodes = mesh->GetNodes();
@@ -189,24 +99,21 @@ Geometry* Geometry::Get(const FiniteElementSpace& fes,
    const int ndofs    = fespace->GetNDofs();
    const kDofQuadMaps* maps = kDofQuadMaps::GetSimplexMaps(*fe, ir);
    NodeCopyByVDim(elements,numDofs,ndofs,dims,geom->eMap,Sx,geom->meshNodes);
-   Geom(dims, numDofs, numQuad, elements,
-        maps->dofToQuadD,
-        geom->meshNodes, geom->J, geom->invJ, geom->detJ);
+   kernels::fem::Geom(dims, numDofs, numQuad, elements,
+                      maps->dofToQuadD,
+                      geom->meshNodes, geom->J, geom->invJ, geom->detJ);
    delete maps;
    return geom;
 }
 
 // *****************************************************************************
-Geometry* Geometry::Get(const FiniteElementSpace& fes,
-                        const IntegrationRule& ir)
+GeometryExtension* GeometryExtension::Get(const FiniteElementSpace& fes,
+                                          const IntegrationRule& ir)
 {
    Mesh *mesh = fes.GetMesh();
    const bool geom_to_allocate = !geom;
-   if (geom_to_allocate) { geom = new Geometry(); }
-   if (!mesh->GetNodes())
-   {
-      // mesh->SetCurvature(1, false, -1, Ordering::byVDIM);
-   }
+   if (geom_to_allocate) { geom = new GeometryExtension(); }
+   mesh->EnsureNodes();
    const GridFunction *nodes = mesh->GetNodes();
 
    const mfem::FiniteElementSpace *fespace = nodes->FESpace();
@@ -246,14 +153,14 @@ Geometry* Geometry::Get(const FiniteElementSpace& fes,
       geom->detJ.allocate(numQuad, elements);
    }
    const kDofQuadMaps* maps = kDofQuadMaps::GetSimplexMaps(*fe, ir);
-   Geom(dims, numDofs, numQuad, elements, maps->dofToQuadD,
-        geom->meshNodes, geom->J, geom->invJ, geom->detJ);
+   kernels::fem::Geom(dims, numDofs, numQuad, elements, maps->dofToQuadD,
+                      geom->meshNodes, geom->J, geom->invJ, geom->detJ);
    delete maps;
    return geom;
 }
 
 // ***************************************************************************
-void Geometry::ReorderByVDim(const GridFunction *nodes)
+void GeometryExtension::ReorderByVDim(const GridFunction *nodes)
 {
    const mfem::FiniteElementSpace *fes = nodes->FESpace();
    const int size = nodes->Size();
@@ -275,7 +182,7 @@ void Geometry::ReorderByVDim(const GridFunction *nodes)
 }
 
 // ***************************************************************************
-void Geometry::ReorderByNodes(const GridFunction *nodes)
+void GeometryExtension::ReorderByNodes(const GridFunction *nodes)
 {
    const mfem::FiniteElementSpace *fes = nodes->FESpace();
    const int size = nodes->Size();
@@ -296,7 +203,4 @@ void Geometry::ReorderByNodes(const GridFunction *nodes)
    delete [] temp;
 }
 
-
-} // namespace geometry
-} // namespace kernels
 } // namespace mfem
