+++ fem/geom_ext.hpp
@@ -9,25 +9,21 @@
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
 
-#ifndef MFEM_FEM_KERNEL_GEOM_HPP
-#define MFEM_FEM_KERNEL_GEOM_HPP
+#ifndef MFEM_FEM_GEOM_HPP
+#define MFEM_FEM_GEOM_HPP
 
-#include "../../config/config.hpp"
-#include "../../general/array_ext.hpp"
+#include "../config/config.hpp"
+#include "../general/array_ext.hpp"
 
-#include "../fespace.hpp"
+#include "fespace.hpp"
 
 namespace mfem
 {
-namespace kernels
-{
-namespace geometry
-{
 
 // ***************************************************************************
-// * Geometry
+// * GeometryExtension
 // ***************************************************************************
-class Geometry
+class GeometryExtension
 {
 public:
    kernels::Array<int> eMap;
@@ -36,17 +32,15 @@ public:
    static const int Jacobian    = (1 << 0);
    static const int JacobianInv = (1 << 1);
    static const int JacobianDet = (1 << 2);
-   static Geometry* Get(const FiniteElementSpace&,
-                        const IntegrationRule&);
-   static Geometry* Get(const FiniteElementSpace&,
-                        const IntegrationRule&,
-                        const Vector&);
+   static GeometryExtension* Get(const FiniteElementSpace&,
+                                 const IntegrationRule&);
+   static GeometryExtension* Get(const FiniteElementSpace&,
+                                 const IntegrationRule&,
+                                 const Vector&);
    static void ReorderByVDim(const GridFunction*);
    static void ReorderByNodes(const GridFunction*);
 };
 
-} // namespace geometry
-} // namespace kernels
 } // namespace mfem
 
 #endif // MFEM_FEM_KERNEL_GEOM_HPP
