+++ fem/kernels/diffusion.cpp
@@ -19,18 +19,159 @@ namespace kernels
 namespace fem
 {
 
+// ****************************************************************************
+// * OCCA 2D Assemble kernel
+// *****************************************************************************
+#ifdef __OCCA__
+static void occaDiffusionAssemble2D(const int NUM_QUAD_1D,
+                                    const int numElements,
+                                    const double* __restrict quadWeights,
+                                    const double* __restrict J,
+                                    const double COEFF,
+                                    double* __restrict oper)
+{
+   const int NUM_QUAD_2D = NUM_QUAD_1D*NUM_QUAD_1D;
+
+   GET_OCCA_CONST_MEMORY(quadWeights);
+   GET_OCCA_CONST_MEMORY(J);
+   GET_OCCA_MEMORY(oper);
+
+   NEW_OCCA_PROPERTY(props);
+   SET_OCCA_PROPERTY(props, NUM_QUAD_1D);
+   SET_OCCA_PROPERTY(props, NUM_QUAD_2D);
+
+   NEW_OCCA_KERNEL(Assemble2D, fem, bidiffusionAssemble.okl, props);
+   Assemble2D(numElements, o_quadWeights, o_J, COEFF, o_oper);
+}
+#endif // __OCCA__
+
+// *****************************************************************************
+// * Diffusion Assemble 2D kernel
+// *****************************************************************************
+static void DiffusionAssemble2D(const int NUM_QUAD_1D,
+                                const int numElements,
+                                const double* __restrict quadWeights,
+                                const double* __restrict J,
+                                const double COEFF,
+                                double* __restrict oper)
+{
+   GET_CONST_PTR(quadWeights);
+   GET_CONST_PTR(J);
+   GET_PTR(oper);
+   const int NUM_QUAD = NUM_QUAD_1D*NUM_QUAD_1D;
+   MFEM_FORALL(e,numElements,
+   {
+      for (int q = 0; q < NUM_QUAD; ++q)
+      {
+         const double J11 = d_J[ijklNM(0,0,q,e,2,NUM_QUAD)];
+         const double J12 = d_J[ijklNM(1,0,q,e,2,NUM_QUAD)];
+         const double J21 = d_J[ijklNM(0,1,q,e,2,NUM_QUAD)];
+         const double J22 = d_J[ijklNM(1,1,q,e,2,NUM_QUAD)];
+         const double c_detJ = d_quadWeights[q] * COEFF / ((J11*J22)-(J21*J12));
+         d_oper[ijkNM(0,q,e,3,NUM_QUAD)] =  c_detJ * (J21*J21 + J22*J22);
+         d_oper[ijkNM(1,q,e,3,NUM_QUAD)] = -c_detJ * (J21*J11 + J22*J12);
+         d_oper[ijkNM(2,q,e,3,NUM_QUAD)] =  c_detJ * (J11*J11 + J12*J12);
+      }
+   });
+}
+
+// *****************************************************************************
+// * Diffusion Assemble 3D kernel
+// *****************************************************************************
+static void DiffusionAssemble3D(const int NUM_QUAD_1D,
+                                const int numElements,
+                                const double* __restrict quadWeights,
+                                const double* __restrict J,
+                                const double COEFF,
+                                double* __restrict oper)
+{
+   GET_CONST_PTR(quadWeights);
+   GET_CONST_PTR(J);
+   GET_PTR(oper);
+   const int NUM_QUAD = NUM_QUAD_1D*NUM_QUAD_1D*NUM_QUAD_1D;
+   MFEM_FORALL(e,numElements,
+   {
+      for (int q = 0; q < NUM_QUAD; ++q)
+      {
+         const double J11 = d_J[ijklNM(0,0,q,e,3,NUM_QUAD)];
+         const double J12 = d_J[ijklNM(1,0,q,e,3,NUM_QUAD)];
+         const double J13 = d_J[ijklNM(2,0,q,e,3,NUM_QUAD)];
+         const double J21 = d_J[ijklNM(0,1,q,e,3,NUM_QUAD)];
+         const double J22 = d_J[ijklNM(1,1,q,e,3,NUM_QUAD)];
+         const double J23 = d_J[ijklNM(2,1,q,e,3,NUM_QUAD)];
+         const double J31 = d_J[ijklNM(0,2,q,e,3,NUM_QUAD)];
+         const double J32 = d_J[ijklNM(1,2,q,e,3,NUM_QUAD)];
+         const double J33 = d_J[ijklNM(2,2,q,e,3,NUM_QUAD)];
+         const double detJ = ((J11 * J22 * J33) + (J12 * J23 * J31) +
+                              (J13 * J21 * J32) - (J13 * J22 * J31) -
+                              (J12 * J21 * J33) - (J11 * J23 * J32));
+         const double c_detJ = d_quadWeights[q] * COEFF / detJ;
+         // adj(J)
+         const double A11 = (J22 * J33) - (J23 * J32);
+         const double A12 = (J23 * J31) - (J21 * J33);
+         const double A13 = (J21 * J32) - (J22 * J31);
+         const double A21 = (J13 * J32) - (J12 * J33);
+         const double A22 = (J11 * J33) - (J13 * J31);
+         const double A23 = (J12 * J31) - (J11 * J32);
+         const double A31 = (J12 * J23) - (J13 * J22);
+         const double A32 = (J13 * J21) - (J11 * J23);
+         const double A33 = (J11 * J22) - (J12 * J21);
+         // adj(J)^Tadj(J)
+         d_oper[ijkNM(0,q,e,6,NUM_QUAD)] = c_detJ *
+            (A11*A11 + A21*A21 + A31*A31); // (1,1)
+         d_oper[ijkNM(1,q,e,6,NUM_QUAD)] = c_detJ *
+            (A11*A12 + A21*A22 + A31*A32); // (1,2), (2,1)
+         d_oper[ijkNM(2,q,e,6,NUM_QUAD)] = c_detJ *
+            (A11*A13 + A21*A23 + A31*A33); // (1,3), (3,1)
+         d_oper[ijkNM(3,q,e,6,NUM_QUAD)] = c_detJ *
+            (A12*A12 + A22*A22 + A32*A32); // (2,2)
+         d_oper[ijkNM(4,q,e,6,NUM_QUAD)] = c_detJ *
+            (A12*A13 + A22*A23 + A32*A33); // (2,3), (3,2)
+         d_oper[ijkNM(5,q,e,6,NUM_QUAD)] = c_detJ *
+            (A13*A13 + A23*A23 + A33*A33); // (3,3)
+      }
+   });
+}
+
+// *****************************************************************************
+void DiffusionAssemble(const int dim,
+                       const int NUM_QUAD_1D,
+                       const int numElements,
+                       const double* __restrict quadWeights,
+                       const double* __restrict J,
+                       const double COEFF,
+                       double* __restrict oper)
+{
+   if (dim==1) { assert(false); }
+   if (dim==2){
+#ifdef __OCCA__
+      if (config::usingOcca()){
+         occaDiffusionAssemble2D(NUM_QUAD_1D, numElements,
+                                 quadWeights, J, COEFF, oper);
+         return;
+      }
+#endif // __OCCA__
+      DiffusionAssemble2D(NUM_QUAD_1D, numElements,
+                          quadWeights, J, COEFF, oper);
+   }
+   if (dim==3) {
+      DiffusionAssemble3D(NUM_QUAD_1D, numElements,
+                          quadWeights, J, COEFF, oper);
+   }
+}
+
 #ifdef __OCCA__
 // *****************************************************************************
-static void oBiPADiffusionMultAdd2D(const int NUM_DOFS_1D,
-                                    const int NUM_QUAD_1D,
-                                    const int numElements,
-                                    const double* __restrict dofToQuad,
-                                    const double* __restrict dofToQuadD,
-                                    const double* __restrict quadToDof,
-                                    const double* __restrict quadToDofD,
-                                    const double* __restrict oper,
-                                    const double* __restrict solIn,
-                                    double* __restrict solOut)
+static void occaDiffusionMultAdd2D(const int NUM_DOFS_1D,
+                                   const int NUM_QUAD_1D,
+                                   const int numElements,
+                                   const double* __restrict dofToQuad,
+                                   const double* __restrict dofToQuadD,
+                                   const double* __restrict quadToDof,
+                                   const double* __restrict quadToDofD,
+                                   const double* __restrict oper,
+                                   const double* __restrict solIn,
+                                   double* __restrict solOut)
 {
    const int NUM_QUAD_2D = NUM_QUAD_1D*NUM_QUAD_1D;
 
@@ -72,14 +213,14 @@ static void oBiPADiffusionMultAdd2D(const int NUM_DOFS_1D,
 // *****************************************************************************
 template<const int NUM_DOFS_1D,
          const int NUM_QUAD_1D> static
-void biPADiffusionMultAdd2D(const int numElements,
-                            const double* __restrict dofToQuad,
-                            const double* __restrict dofToQuadD,
-                            const double* __restrict quadToDof,
-                            const double* __restrict quadToDofD,
-                            const double* __restrict oper,
-                            const double* __restrict solIn,
-                            double* __restrict solOut)
+void DiffusionMultAssembled2D(const int numElements,
+                              const double* __restrict dofToQuad,
+                              const double* __restrict dofToQuadD,
+                              const double* __restrict quadToDof,
+                              const double* __restrict quadToDofD,
+                              const double* __restrict oper,
+                              const double* __restrict solIn,
+                              double* __restrict solOut)
 {
    const int NUM_QUAD = NUM_QUAD_1D*NUM_QUAD_1D;
    MFEM_FORALL(e, numElements,
@@ -183,14 +324,14 @@ void biPADiffusionMultAdd2D(const int numElements,
 // *****************************************************************************
 template<const int NUM_DOFS_1D,
          const int NUM_QUAD_1D> static
-void biPADiffusionMultAdd3D(const int numElements,
-                            const double* __restrict dofToQuad,
-                            const double* __restrict dofToQuadD,
-                            const double* __restrict quadToDof,
-                            const double* __restrict quadToDofD,
-                            const double* __restrict oper,
-                            const double* __restrict solIn,
-                            double* __restrict solOut)
+void DiffusionMultAssembled3D(const int numElements,
+                              const double* __restrict dofToQuad,
+                              const double* __restrict dofToQuadD,
+                              const double* __restrict quadToDof,
+                              const double* __restrict quadToDofD,
+                              const double* __restrict oper,
+                              const double* __restrict solIn,
+                              double* __restrict solOut)
 {
    const int NUM_QUAD = NUM_QUAD_1D*NUM_QUAD_1D*NUM_QUAD_1D;
    MFEM_FORALL(e, numElements,
@@ -374,27 +515,27 @@ typedef void (*fDiffusionMultAdd)(const int numElements,
                                   double* __restrict solOut);
 
 // *****************************************************************************
-void biPADiffusionMultAdd(const int DIM,
-                          const int NUM_DOFS_1D,
-                          const int NUM_QUAD_1D,
-                          const int numElements,
-                          const double* __restrict dofToQuad,
-                          const double* __restrict dofToQuadD,
-                          const double* __restrict quadToDof,
-                          const double* __restrict quadToDofD,
-                          const double* __restrict op,
-                          const double* __restrict x,
-                          double* __restrict y)
+void DiffusionMultAssembled(const int DIM,
+                            const int NUM_DOFS_1D,
+                            const int NUM_QUAD_1D,
+                            const int numElements,
+                            const double* __restrict dofToQuad,
+                            const double* __restrict dofToQuadD,
+                            const double* __restrict quadToDof,
+                            const double* __restrict quadToDofD,
+                            const double* __restrict op,
+                            const double* __restrict x,
+                            double* __restrict y)
 {
 
 #ifdef __OCCA__
    if (config::usingOcca()){
       assert(DIM==2);
-      oBiPADiffusionMultAdd2D(NUM_DOFS_1D, NUM_QUAD_1D,
-                              numElements,
-                              dofToQuad, dofToQuadD,
-                              quadToDof, quadToDofD,
-                              op, x, y);
+      occaDiffusionMultAssembled2D(NUM_DOFS_1D, NUM_QUAD_1D,
+                                   numElements,
+                                   dofToQuad, dofToQuadD,
+                                   quadToDof, quadToDofD,
+                                   op, x, y);
       return;
    }
 #endif // __OCCA__
@@ -404,43 +545,43 @@ void biPADiffusionMultAdd(const int DIM,
    assert(LOG2(NUM_QUAD_1D)<=8);
    static std::unordered_map<unsigned int, fDiffusionMultAdd> call =
    {
-      {0x20101,&biPADiffusionMultAdd2D<1,1>},
-      {0x20201,&biPADiffusionMultAdd2D<2,1>},
-      {0x20202,&biPADiffusionMultAdd2D<2,2>},
-      {0x20303,&biPADiffusionMultAdd2D<3,3>},
-      {0x20404,&biPADiffusionMultAdd2D<4,4>},
-      {0x20505,&biPADiffusionMultAdd2D<5,5>},
-      {0x20606,&biPADiffusionMultAdd2D<6,6>},
-      {0x20707,&biPADiffusionMultAdd2D<7,7>},
-      {0x20808,&biPADiffusionMultAdd2D<8,8>},/*
-      {0x20909,&biPADiffusionMultAdd2D<9,9>},
-      {0x20A0A,&biPADiffusionMultAdd2D<10,10>},
-      {0x20B0B,&biPADiffusionMultAdd2D<11,11>},
-      {0x20C0C,&biPADiffusionMultAdd2D<12,12>},
-      {0x20D0D,&biPADiffusionMultAdd2D<13,13>},
-      {0x20E0E,&biPADiffusionMultAdd2D<14,14>},
-      {0x20F0F,&biPADiffusionMultAdd2D<15,15>},
-      {0x21010,&biPADiffusionMultAdd2D<16,16>},
-      {0x21111,&biPADiffusionMultAdd2D<17,17>},*/
+      {0x20101,&DiffusionMultAssembled2D<1,1>},
+      {0x20201,&DiffusionMultAssembled2D<2,1>},
+      {0x20202,&DiffusionMultAssembled2D<2,2>},
+      {0x20303,&DiffusionMultAssembled2D<3,3>},
+      {0x20404,&DiffusionMultAssembled2D<4,4>},
+      {0x20505,&DiffusionMultAssembled2D<5,5>},
+      {0x20606,&DiffusionMultAssembled2D<6,6>},
+      {0x20707,&DiffusionMultAssembled2D<7,7>},
+      {0x20808,&DiffusionMultAssembled2D<8,8>},/*
+      {0x20909,&DiffusionMultAssembled2D<9,9>},
+      {0x20A0A,&DiffusionMultAssembled2D<10,10>},
+      {0x20B0B,&DiffusionMultAssembled2D<11,11>},
+      {0x20C0C,&DiffusionMultAssembled2D<12,12>},
+      {0x20D0D,&DiffusionMultAssembled2D<13,13>},
+      {0x20E0E,&DiffusionMultAssembled2D<14,14>},
+      {0x20F0F,&DiffusionMultAssembled2D<15,15>},
+      {0x21010,&DiffusionMultAssembled2D<16,16>},
+      {0x21111,&DiffusionMultAssembled2D<17,17>},*/
 
-      {0x30101,&biPADiffusionMultAdd3D<1,1>},
-      {0x30201,&biPADiffusionMultAdd3D<2,1>},
-      {0x30202,&biPADiffusionMultAdd3D<2,2>},
-      {0x30203,&biPADiffusionMultAdd3D<2,3>},
-      {0x30303,&biPADiffusionMultAdd3D<3,3>},
-      {0x30404,&biPADiffusionMultAdd3D<4,4>},
-      {0x30505,&biPADiffusionMultAdd3D<5,5>},
-      {0x30606,&biPADiffusionMultAdd3D<6,6>},
-      {0x30707,&biPADiffusionMultAdd3D<7,7>},
-      {0x30808,&biPADiffusionMultAdd3D<8,8>},/*
-      {0x30909,&biPADiffusionMultAdd3D<9,9>},
-      {0x30A0A,&biPADiffusionMultAdd3D<10,10>},
-      {0x30B0B,&biPADiffusionMultAdd3D<11,11>},
-      {0x30C0C,&biPADiffusionMultAdd3D<12,12>},
-      {0x30D0D,&biPADiffusionMultAdd3D<13,13>},
-      {0x30E0E,&biPADiffusionMultAdd3D<14,14>},
-      {0x30F0F,&biPADiffusionMultAdd3D<15,15>},
-      {0x31010,&biPADiffusionMultAdd3D<16,16>},*/
+      {0x30101,&DiffusionMultAssembled3D<1,1>},
+      {0x30201,&DiffusionMultAssembled3D<2,1>},
+      {0x30202,&DiffusionMultAssembled3D<2,2>},
+      {0x30203,&DiffusionMultAssembled3D<2,3>},
+      {0x30303,&DiffusionMultAssembled3D<3,3>},
+      {0x30404,&DiffusionMultAssembled3D<4,4>},
+      {0x30505,&DiffusionMultAssembled3D<5,5>},
+      {0x30606,&DiffusionMultAssembled3D<6,6>},
+      {0x30707,&DiffusionMultAssembled3D<7,7>},
+      {0x30808,&DiffusionMultAssembled3D<8,8>},/*
+      {0x30909,&DiffusionMultAssembled3D<9,9>},
+      {0x30A0A,&DiffusionMultAssembled3D<10,10>},
+      {0x30B0B,&DiffusionMultAssembled3D<11,11>},
+      {0x30C0C,&DiffusionMultAssembled3D<12,12>},
+      {0x30D0D,&DiffusionMultAssembled3D<13,13>},
+      {0x30E0E,&DiffusionMultAssembled3D<14,14>},
+      {0x30F0F,&DiffusionMultAssembled3D<15,15>},
+      {0x31010,&DiffusionMultAssembled3D<16,16>},*/
    };
    if (!call[id])
    {
@@ -462,7 +603,6 @@ void biPADiffusionMultAdd(const int DIM,
             d_op, d_x, d_y);
 }
 
-// *****************************************************************************
 } // namespace fem
 } // namespace kernels
 } // namespace mfem
